Addr class for sock address functions
In most cases, usable in place of the old because it inherits from
tuple, except that it's not immutable so it can't be used as a
dictionary key (I think?)
This makes the sock parameter passed to Servers unnecessary.
I also moved the Info class definition for no clear reason.