Don't forget a server that fails an infoResponse
This fixes a bug that occurred when two heartbeats arrived together, as
sometimes happens: two getinfo responses would be sent, and the
infoResponse with the first challenge would fail to match the second
challenge, and be rejected: the second infoResponse would then be
ignored. Now the first infoResponse still fails but the second will
succeed; servers that fail entirely will only be deleted when they time
out.